refactor: Refactor code

+++ src/com/samin/Q7/FloatCalcError.java
@@ -10,8 +10,9 @@ import java.math.BigDecimal;
  */
 public class FloatCalcError {
     public static void main(String[] args) {
+        // 小数默认为double类型，所以声明float类型的变量需要后面加'f'或'F'，或者使用(float)num转换类型
         float a = 68399.22f;
-        float b = 124052.96f;
+        float b = (float) 124052.96;
         System.out.println(a + b); // 出现精度问题，结果错误
         BigDecimal bigDecimal1 = new BigDecimal(Float.toString(a));
         BigDecimal bigDecimal2 = new BigDecimal(Float.toString(b));
